requirements.txt
fastapi>=0.116.0  # for API (Python 3.12 compatible)
GPUtil==1.4.0  # optional
h5py>=3.12.1
imagehash==4.2.0  # optional
matplotlib>=3.8.0
networkx>=3.1  # optional
numpy>=1.26
onnx>=1.16.0,<2
orjson>=3.9.0  # fast JSON parsing for metadata cache (5-10x faster than stdlib)
# Use the GPU build of ONNX Runtime; do not install CPU + GPU together
onnxruntime-gpu>=1.22.0
onnxsim>=0.4.36
opencv-python>=4.8
pandas>=2.1
pillow>=10.0.0
pyarrow>=14.0.0  # Arrow IPC for zero-copy metadata cache (memory-mapped, shared across workers)
prometheus-client>=0.15.0  # optional
psutil>=5.9.0
python-multipart>=0.0.9  # for API
pyyaml>=6.0.1
requests>=2.31
protobuf>=4.25.3,<6
werkzeug>=2.3,<5
scikit-learn>=1.4
scipy>=1.11
safetensors>=0.4.2
seaborn>=0.13.2
tensorboard>=2.14,<2.21
bitsandbytes>=0.41.0  # for 8-bit AdamW optimizer (memory efficient)
# Upgrade PyTorch stack for modern APIs and CUDA 12.8+ / SM_120 support
torch>=2.8.0
torchvision>=0.23.0
tqdm>=4.66.4
uvicorn>=0.35.0  # for API
wandb>=0.16.4  # optional
wordcloud>=1.9.3  # optional

# Triton for torch.compile optimization (15-35% training speedup)
triton-windows>=3.4.0,<3.5

# New training stack dependencies
pytorch-lightning>=2.5.3
torchmetrics>=1.3.0
ruamel.yaml>=0.17.31  # for config loading/writing
